Refactor client.ts to use decorators to attach config
