+++ src/API.ts
@@ -1,6 +1,6 @@
 import superagent from 'superagent';
 import Cookies from 'universal-cookie';
-import { addHeadersFactory } from './utils';
+import { addHeadersFactory } from './utils/api';
 import type request from 'superagent';
 import type { PloneClientConfig } from './client';
